Don't raise in ReponseTemplate#compile when nil mediatype
When a response_template doesn't detect a media type, an exception is raised which results in the below error when running "rake praxis:routes":

"Praxis::Exceptions::InvalidConfiguration: Could not default :media_type argument for response template ok. Resource V1::Resources::Posts does not have an associated mediatype and none was passed"

This breaks the procedure in "Getting Started".

Fixes #28

Signed-off-by: Justin Gaylor justin@rightscale.com