Remove `actions` and `action` methods from controller concern.
These methods return definition actions anyway, so it seems clearer
to have the code follow up the definition object. That is: 
* ```ctrl.actions``` should now be ```ctrl.definition.actions```
* ```ctrl.action(:name)``` should now be ```ctrl.definition.actions[:name]```

This opens the door to define both design and controller in a single file/class
for quick tests.
Adapted the small parts of the code to access the actions the new way.

Signed-off-by: Josep M. Blanquer <blanquer@rightscale.com>